refactor(card): one store and theme lifecycle for the card and the panel
`haventory-card` and `haventory-panel` are the two elements Home
Assistant instantiates itself, and each carried the same five members:
the `hass` setter that sets the language, builds the store once and
watches it, the re-subscribe on reconnect, the teardown, and the
`color-scheme` publish. They now share `StoreHostElement`; each keeps
what it renders and what configuration it reads.

The panel's `requestUpdate()` after the store is built is what the
shared setter does — on the card it lands one render earlier than
before and changes nothing on screen.

Refs #231
